requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
openpyxl==3.1.2
ics==0.7.2
python-telegram-bot==20.7
pytz==2023.3
